Added utility classes (MathUtils, RandomUtils, Point, Direction)
The MathUtils class contains some trigonometric methods (sin, cos, atan2) and methods for computing the square root and square of a number. The data type for numbers is float and angles are in degrees.

The RandomUtils class contains methods for getting random numbers in a specified range and random booleans.

Point and Direction are just containers for two floats with some added functionality.