Added a ServerGame and ClientGame class
The ServerGame and ClientGame classes are reponsible for handling the logic of the server and the client. They setup the server and client and set message handlers for processing the messages.

They are both stored in the JamKartApp class (previously named ClientApp) and get stopped when the window is closed.

They client sends a message to the server when joining with the name and the server currently only prints this name.

The games are started by the scenes. The SceneJoinGame starts a ClientGame and the SceneCreateGame starts a ServerGame and a ClientGame.

There is also now a icon for the game in the resources directory.