Added basic input handling and car movement
Added a KeyEventHandler class that is responsible for receiving key events and storing in booleans if a key is currently pressed.

The KeyEventHandler can be accessed by the Car and is used to control it. A, D rotate the car left and right, W moves it forward, S moves it (half as fast) backward and space brakes the car.