Added road
Added a Road class. There is one instance of Road per Map. The actual road is represented in the Road class as a Polygon. This enables us to call the contains method of the Polygon class to determine whether the car is on the road. The Road is constructed by calling the addPoint method and it is drawn by calling the draw method.