Added support for lap counting and cleaned up some code
Added a RoadPathTracker that counts the number of laps passed. It does this by counting the number of RoadPathMarkers the car passes. A RoadPathMarker is placed on the road and must be passed in order to finish the lap. This should make it impossible to do crazy shortcuts.

Also moved some duplicate drawing code for drawing rotated objects to a DrawingUtils class.