Created superclass for cars
Renamed ClientCar to ClientMyCar and created a superclass for ClientMyCar and ClientOtherCar called ClientCar (not to be confused with the old name of ClientMyCar). All shared code of the two classes is now in this class.

Added a boolean "finished" and a float "finishTime" to the ClientCar class. When a car finishes, the method "finish" is called and the values are set. This will later be used to create a leaderboard.